requirements.txt
python-docx
docx2txt
PyMuPDF
Pillow
openai
fastapi
uvicorn
python-dotenv